+++ .build/tasks/release.module.build.ps1
@@ -34,7 +34,10 @@ param (
     $PSModuleFeed = (property PSModuleFeed 'PSGallery'),
 
     [Parameter()]
-    $SkipPublish = (property SkipPublish '')
+    $SkipPublish = (property SkipPublish ''),
+
+    [Parameter()]
+    $WhatIf = (property WhatIf '')
 )
 
 # Synopsis: Create ReleaseNotes from changelog and update the Changelog for release
@@ -229,7 +232,15 @@ task package_module_nupkg {
                 $Prerelease = "-" + $Prerelease
             }
             Write-Build Yellow ("  Packaging Required Module {0} v{1}{2}" -f $Module.Name, $Module.Version.ToString(), $Prerelease)
-            Publish-Module -Repository output -ErrorAction SilentlyContinue -Path $module.ModuleBase
+
+            if ($WhatIf)
+            {
+                Publish-Module -Repository output -ErrorAction SilentlyContinue -Path $module.ModuleBase -WhatIf
+            }
+            else
+            {
+                Publish-Module -Repository output -ErrorAction SilentlyContinue -Path $module.ModuleBase
+            }
         }
     }
 
@@ -241,7 +252,15 @@ task package_module_nupkg {
         Force           = $true
     }
 
-    Publish-Module @PublishModuleParams
+    if ($WhatIf)
+    {
+        Publish-Module @PublishModuleParams -WhatIf
+    }
+    else
+    {
+        Publish-Module @PublishModuleParams
+    }
+
     Write-Build Green "`n  Packaged $ProjectName NuGet package `n"
     Write-Build DarkGray "  Cleaning up"
 
@@ -289,7 +308,11 @@ task publish_module_to_gallery -if ($GalleryApiToken -and (Get-Command -Name 'Pu
         ReleaseNotes    = $releaseNotesForLatestRelease
     }
 
-    if (!$SkipPublish)
+    if ($WhatIf)
+    {
+        Publish-Module @PublishModuleParams -WhatIf
+    }
+    elseif (!$SkipPublish)
     {
         Publish-Module @PublishModuleParams
     }